+++ common/src/main/java/com/diontryban/transparent/mixin/client/HumanoidArmorLayerMixin.java
@@ -1,0 +1,33 @@
+package com.diontryban.transparent.mixin.client;
+
+import com.diontryban.transparent.Transparent;
+import com.diontryban.transparent.client.render.TransparentRenderTypes;
+import net.minecraft.client.model.HumanoidModel;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.entity.RenderLayerParent;
+import net.minecraft.client.renderer.entity.layers.HumanoidArmorLayer;
+import net.minecraft.client.renderer.entity.layers.RenderLayer;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.LivingEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(HumanoidArmorLayer.class)
+public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends HumanoidModel<T>, A extends HumanoidModel<T>> extends RenderLayer<T, M> {
+    public HumanoidArmorLayerMixin(RenderLayerParent<T, M> renderer) {
+        super(renderer);
+    }
+
+    // The Minecraft Development IntelliJ Plugin claims this is incorrect, but it is.
+    // On Fabric, there's only one renderModel method, but on NeoForge there are two, so it has to be specified.
+    @Redirect(
+            method = "renderModel(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/client/model/Model;ILnet/minecraft/resources/ResourceLocation;)V",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/RenderType;armorCutoutNoCull(Lnet/minecraft/resources/ResourceLocation;)Lnet/minecraft/client/renderer/RenderType;")
+    )
+    public RenderType redirectArmorCutoutNoCull(ResourceLocation location) {
+        return Transparent.CONFIG.playerArmor ?
+                TransparentRenderTypes.armorCutoutNoCull(location) :
+                RenderType.armorCutoutNoCull(location);
+    }
+}